list: support multiple labels on the same oid.
PKCS#11 can create many keys reusing the same label and all pointing to
the same NXP SE050 key.

This commit enhances the --list-objects option displaying all the
different labels.

It is recommended to dump its output to a file to avoid wrapping lines
whenever many labels are associated to the same key

ie:
./fio-se05x-cli --list-objects --token name --pin 1111 2> /tmp/file

Signed-off-by: Jorge Ramirez-Ortiz <jorge@foundries.io>